Move all users private function to users_helper.py
