+++ accountlockout/utils.py
@@ -11,6 +11,7 @@ from django.utils.encoding import force_bytes
 from django.contrib.auth.tokens import default_token_generator
 
 import accountlockout.helper.ip_helper
+import accountlockout.helper.users_helper
 from core.models import User, SiteConfiguration
 from defender.connection import get_redis_connection
 from defender.data import store_login_attempt
@@ -24,7 +25,7 @@ REDIS_SERVER = get_redis_connection()
 def is_already_locked(request, get_username=None, username=None):
     """Parse the username & IP from the request, and see if it's
     already locked."""
-    user_blocked = users.__is_already_locked(request, username, get_username)
+    user_blocked = accountlockout.helper.users_helper.__is_already_locked(request, username, get_username)
     ip_blocked = accountlockout.helper.ip_helper.__is_source_already_locked(accountlockout.helper.ip_helper.__get(request))
 
     if def_config.LOCKOUT_BY_IP_USERNAME:
@@ -65,7 +66,7 @@ def get_attemps_left(request):
 
 def lockout_response(request):
     """ if we are locked out, here is the response """
-    username = users.get_username_from_request(request)
+    username = accountlockout.helper.users_helper.get_username_from_request(request)
     if def_config.LOCKOUT_TEMPLATE:
         context = {
             'cooloff_time_seconds': def_config.COOLOFF_TIME,
@@ -94,7 +95,7 @@ def add_login_attempt_to_db(request, login_valid,get_usernamefunc=None,username=
         # If we don't want to store in the database, then don't proceed.
         return
 
-    username = username or users.get_username(request, get_usernamefunc)
+    username = username or accountlockout.helper.users_helper.__get_username(request, get_usernamefunc)
 
     user_agent = request.META.get('HTTP_USER_AGENT', '<unknown>')[:255]
     ip_address = accountlockout.helper.ip_helper.__get(request)
